requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==3.2.9
django-common-helpers==0.9.2
django-crontab==0.7.1
django-debug-toolbar==3.2.2
django-restframework==0.0.1
djangorestframework==3.12.4
django-mysql==3.8.1
flake8==4.0.1
kombu==5.2.2
mccabe==0.6.1
mysqlclient==2.1.0
mysql-connector-python==8.0.26
prompt-toolkit==3.0.23
protobuf==3.19.1
pycodestyle==2.8.0
pyflakes==2.4.0
pytz==2021.3
redis==4.0.2
six==1.16.0
sqlparse==0.4.2
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
